+++ packages/vitePress/docs/.vitepress/Theme/index.ts
@@ -3,7 +3,7 @@ import DefaultTheme from 'vitepress/theme';
 import PreView from "./components/PreView.vue";
 import Home from './components/Home.vue'
 import { type EnhanceAppContext } from 'vitepress'
-
+import './style/var.css'
 export default {
   ...DefaultTheme,
   enhanceApp({ app }:EnhanceAppContext) {
