+++ packages/vitePress/docs/.vitepress/Theme/style/var.css
@@ -1,0 +1,9 @@
+:root {
+    /* 标题 */
+    --vp-home-hero-name-color: transparent;
+    --vp-home-hero-name-background: linear-gradient( 135deg, #a8fa0f 10%, #0becda 100%);
+    
+    /* 图标背景 */
+    --vp-home-hero-image-background-image: linear-gradient( 135deg, #a8fa0f 10%, #0becda 100%);
+    --vp-home-hero-image-filter: blur(150px);
+  }
