+++ packages/vitePress/docs/guide/change-log.md
@@ -1,4 +1,3 @@
-
 ---
 layout: doc
 ---
