--- packages/vitePress/docs/public/json/contributors.json
@@ -1,50 +1,0 @@
-[
-    {
-        "avatar": "nice-knight",
-        "name": "nice-knight",
-        "github": "nice-knight",
-        "twitter": "nice-knight",
-        "description": "是多个开源哭的贡献者，活动于各个社区 微信号110",
-        "packages": ["motion", "gesture", "sound"]
-      },
-      {
-        "avatar": "kitymore",
-        "name": "kitymore",
-        "github": "kitymore",
-        "twitter": "kitymore",
-        "description": "是多个开源哭的贡献者，活动于各个社区 微信号1190.",
-        "packages": ["components"]
-      },
-      {
-        "avatar": "lqh1008",
-        "name": "lqh1008",
-        "github": "lqh1008",
-        "twitter": "lqh1008",
-        "description": "是多个开源哭的贡献者，活动于各个社区 微信号1910.",
-        "packages": ["components"]
-      },
-      {
-        "avatar": "Patrick-CHenYang",
-        "name": "Patrick-CHenYang",
-        "github": "Patrick-CHenYang",
-        "twitter": "Patrick-CHenYang",
-        "description": "是多个开源哭的贡献者，活动于各个社区 微信号1910",
-        "packages": ["components"]
-      },
-      {
-        "avatar": "fushengcf",
-        "name": "fushengcf",
-        "github": "fushengcf",
-        "twitter": "fushengcf",
-        "description": "是多个开源哭的贡献者，活动于各个社区 微信号1990",
-        "packages": ["components"]
-      },
-      {
-        "avatar": "macx-cxxi",
-        "name": "macx-cxxi",
-        "github": "macx-cxxi",
-        "twitter": "macx-cxxi",
-        "description": "是多个开源哭的贡献者，活动于各个社区 微信号1990",
-        "packages": ["components"]
-      }
-]
